Deleted duplicated code
