+++ mrpt_pointcloud_pipeline/CMakeLists.txt
@@ -42,7 +42,7 @@ ENDIF()
 find_package(mp2p_icp_filters REQUIRED)
 
 add_executable(${PROJECT_NAME}_node
-              src/mrpt_pointcloud_pipeline_node.cpp
+              src/main.cpp
               include/mrpt_pointcloud_pipeline/mrpt_pointcloud_pipeline_node.h)
 
 target_include_directories(${PROJECT_NAME}_node
